+++ src/swap/defi/uni-v2-based/plugins/spookySwap.ts
@@ -168,7 +168,7 @@ export function makeSpookySwapPlugin(
       savedAction: {
         actionType: 'swap',
         swapInfo,
-        isEstimate: false,
+        isEstimate: true,
         toAsset: {
           pluginId: request.toWallet.currencyInfo.pluginId,
           tokenId: request.toTokenId,