+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- changed: Standardize DEX quotes to 5% slippage and estimate quote
+
 ## 2.4.3 (2024-06-03)
 
 - added: (Exolix) Add Piratechain